Fixed unit tests and changed functionality
