+++ src/components/Tabs/Tabs.test.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Tabs, mapDispatchToProps } from "./Tabs";
+import Tabs from "./Tabs";
 import TabPane from "../TabPane/TabPane";
 import {
   render,
@@ -36,16 +36,6 @@ const childProps = {
   facId: testFacId,
   checkedOutBy: testUserId,
 };
-const mockDispatch = jest.fn();
-jest.mock('react-redux', () => ({
-  useSelector: jest.fn(),
-  useDispatch: () => mockDispatch,
-  connect: (mapStateToProps, mapDispatchToProps) => (ReactComponent) => ({
-    mapStateToProps,
-    mapDispatchToProps,
-    ReactComponent,
-  }),
-}));
 
 jest.mock("../../utils/api/monitoringPlansApi", () => {
   return {
@@ -81,7 +71,10 @@ const TabsUsage = (bool) => (
     removeTabs={jest.fn()}
     checkedOutLocations={testCheckedOutLocations}
     user={{ firstName: testFirstName, userId: testUserId }}
+    setCheckout={jest.fn()}
     workspaceSection={MONITORING_PLAN_STORE_NAME}
+    setCurrentTabIndex={jest.fn()}
+    currentTabIndex={0}
   >
     <TabPane {...childProps}>Tab1 Content</TabPane>
     <TabPane title="Select configurations">Tab2 Content</TabPane>
@@ -113,15 +106,12 @@ describe("testing a reusable Tabs component", () => {
   });
   test("renders the specified initial tabpane content ", () => {
     render(<TabsUsage />);
-    const initTabContent = screen.getAllByText("Tab");
-    expect(initTabContent[0]).not.toBeUndefined();
+    const initTabContent = screen.getByText("Tab1 Content");
+    expect(initTabContent).not.toBeUndefined();
   });
   test("renders the user selected tab", async () => {
-    const dispatch = jest.fn();
-    const actionProps = mapDispatchToProps(dispatch);
 
     // verify the appropriate action was called
-    actionProps.setCheckout();
     const { container } = await waitForElement(() =>
       render(
         <Tabs
@@ -132,6 +122,8 @@ describe("testing a reusable Tabs component", () => {
           user={{ firstName: testFirstName }}
           setCheckout={jest.fn()}
           workspaceSection={MONITORING_PLAN_STORE_NAME}
+          setCurrentTabIndex={jest.fn()}
+          currentTabIndex={0}
         >
           <TabPane {...childProps}>Tab1 Content</TabPane>
           <TabPane title="Select configurations">Tab2 Content</TabPane>
